Some casts

+++ Source/Engine/ECS/Animation/Animation.cpp
@@ -10,28 +10,25 @@
 #include <assimp/scene.h>
 #include <assimp/postprocess.h>
 
+// ----------------------------------------------------
+// 									PUBLIC
+// ----------------------------------------------------
 
-
-// ---------------------------------------------------- 
-// 									PUBLIC														
-// ---------------------------------------------------- 
-
-Animation::Animation(const SkeletalMesh& skeleton, const fs::path& relative) :
-	bonesAnimKeys{},
-	nrKeys{ 0 },
-	duration{ 0 },
-	ticksPerSecond{ 0 },
-	id{}
+Animation::Animation(const SkeletalMesh &skeleton, const fs::path &relative) : bonesAnimKeys{},
+																																							 nrKeys{0},
+																																							 duration{0},
+																																							 ticksPerSecond{0},
+																																							 id{}
 {
 	// E.g. relative = "Mutant/Drunk_Walk/anim.gltf"
 	// E.g. absolute = "D:GameEngine/Assets/Models/Skeletal/Mutant/Drunk_Walk/anim.gltf"
 	fs::path absolute = (Paths::GetSkeletalModelsPath() / relative);
 
 	Assimp::Importer importer;
-	const aiScene* scene = importer.ReadFile(absolute.string(), 
+	const aiScene *scene = importer.ReadFile(absolute.string(),
 																					 aiProcess_Triangulate |
-																					 aiProcess_LimitBoneWeights |
-																					 aiProcess_JoinIdenticalVertices);
+																							 aiProcess_LimitBoneWeights |
+																							 aiProcess_JoinIdenticalVertices);
 	if (!scene || !scene->mRootNode)
 	{
 		CONSOLE_ERROR("Assimp importer error: {}", importer.GetErrorString());
@@ -43,39 +40,39 @@ Animation::Animation(const SkeletalMesh& skeleton, const fs::path& relative) :
 		return;
 	}
 
-	aiAnimation* animation = scene->mAnimations[0];
-	duration = animation->mDuration;
-	ticksPerSecond = animation->mTicksPerSecond;
+	aiAnimation *animation = scene->mAnimations[0];
+	duration = static_cast<f32>(animation->mDuration);
+	ticksPerSecond = static_cast<f32>(animation->mTicksPerSecond);
 
 	bonesAnimKeys = std::make_unique<BoneAnimationKeys[]>(skeleton.nrBones);
 	nrKeys = skeleton.nrBones;
-	
+
 	for (u32 i = 0; i < animation->mNumChannels; i++)
 	{
-		aiNodeAnim* channel = animation->mChannels[i];
-		const char* channelName = channel->mNodeName.C_Str();
+		aiNodeAnim *channel = animation->mChannels[i];
+		const char *channelName = channel->mNodeName.C_Str();
 
 		i32 boneIndex = skeleton.FindBone(channelName);
 		if (boneIndex == -1)
 			continue;
 
-		auto& boneKeys = bonesAnimKeys[boneIndex];
+		auto &boneKeys = bonesAnimKeys[boneIndex];
 		LoadBoneKeys(boneKeys, channel);
 	}
 }
 
 // ----------------------------------------------------
-//										PRIVATE													
+//										PRIVATE
 // ----------------------------------------------------
 
-void Animation::LoadBoneKeys(BoneAnimationKeys& boneKeys, const aiNodeAnim* channel)
+void Animation::LoadBoneKeys(BoneAnimationKeys &boneKeys, const aiNodeAnim *channel)
 {
 	boneKeys.posKeys = std::make_unique<KeyPosition[]>(channel->mNumPositionKeys);
 	boneKeys.nrPosKeys = channel->mNumPositionKeys;
 	for (u32 i = 0; i < channel->mNumPositionKeys; i++)
 	{
-		aiVectorKey& tmp = channel->mPositionKeys[i];
-		boneKeys.posKeys[i].timeStamp = tmp.mTime;
+		aiVectorKey &tmp = channel->mPositionKeys[i];
+		boneKeys.posKeys[i].timeStamp = static_cast<f32>(tmp.mTime);
 		boneKeys.posKeys[i].position = vec3f(tmp.mValue.x, tmp.mValue.y, tmp.mValue.z);
 	}
 
@@ -83,8 +80,8 @@ void Animation::LoadBoneKeys(BoneAnimationKeys& boneKeys, const aiNodeAnim* chan
 	boneKeys.nrRotKeys = channel->mNumRotationKeys;
 	for (u32 i = 0; i < channel->mNumRotationKeys; i++)
 	{
-		aiQuatKey& tmp = channel->mRotationKeys[i];
-		boneKeys.rotKeys[i].timeStamp = tmp.mTime;
+		aiQuatKey &tmp = channel->mRotationKeys[i];
+		boneKeys.rotKeys[i].timeStamp = static_cast<f32>(tmp.mTime);
 		boneKeys.rotKeys[i].orientation = quat(tmp.mValue.w, tmp.mValue.x, tmp.mValue.y, tmp.mValue.z);
 	}
 
@@ -92,8 +89,8 @@ void Animation::LoadBoneKeys(BoneAnimationKeys& boneKeys, const aiNodeAnim* chan
 	boneKeys.nrScaleKeys = channel->mNumScalingKeys;
 	for (u32 i = 0; i < channel->mNumScalingKeys; i++)
 	{
-		aiVectorKey& tmp = channel->mScalingKeys[i];
-		boneKeys.scaleKeys[i].timeStamp = tmp.mTime;
+		aiVectorKey &tmp = channel->mScalingKeys[i];
+		boneKeys.scaleKeys[i].timeStamp = static_cast<f32>(tmp.mTime);
 		boneKeys.scaleKeys[i].scale = vec3f(tmp.mValue.x, tmp.mValue.y, tmp.mValue.z);
 	}
 }
